feat(extensions): add DelayedAsync method and refactor TimeProvider extensions
